requirements.txt
langchain==0.2.14
langchain_community==0.2.12
langchain_openai==0.1.21
openai==1.40.6
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.37.1
